LayoutTests/imported/w3c/web-platform-tests/webxr/layers/xrWebGLBinding_constructor.https-expected.txt


FAIL Ensure that XRWebGLBinding's constructor throws appropriate errors using webgl assert_equals: Should get InvalidStateError for creating with inline session. expected "InvalidStateError" but got "ReferenceError"
FAIL Ensure that XRWebGLBinding's constructor throws appropriate errors using webgl2 assert_equals: Should get InvalidStateError for creating with inline session. expected "InvalidStateError" but got "ReferenceError"
